Updated ReadMe
Updated readme so that I can begin to organize my GitHub page and make it easier to navigate and meaningful. This is the first step to making sure that what I provide or show employers or clients is something that will be beneficial. I will add more to this once I have a local branch on my pc and have had more time to gather details.